createforcast screen

+++ domain/forcast/src/main/java/com/example/forcast/models/Coord.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Coord(
     @SerialName("lat")
-    val lat: Int,
+    val lat: Double,
     @SerialName("lon")
     val lon: Double
 )
